Move tests to AvroSuite
